+++ src/pages/login/components/GoogleLoginButton.tsx
@@ -1,9 +1,9 @@
-import { getClubsInfo } from "@api/club/getClubsInfo";
-import { fetchLoginData } from "@api/login/fetchLoginData";
 import Button from "@components/Button";
 import { auth } from "@config/firebaseConfig";
 import { useAuth } from "@contexts/AuthContext";
 import usePrefixedNavigate from "@hooks/usePrefixedNavigate";
+import { fetchLoginData } from "@libs/api/auth";
+import { getClubsInfo } from "@libs/api/club";
 import ROUTE from "@libs/constant/path";
 import { GoogleAuthProvider, signInWithPopup } from "firebase/auth";
