+++ src/types/auth.ts
@@ -1,0 +1,6 @@
+type LoginData = {
+  accessToken: string;
+  refreshToken: string;
+};
+
+export type { LoginData };
