+++ src/types/group.ts
@@ -1,0 +1,13 @@
+type GroupInfoProps = {
+  clubId: number;
+};
+
+type GroupInfoResponseData = {
+  groupId: number;
+  groupName: string;
+  groupLink: string;
+  groupDescription: string;
+  groupAmount: number;
+};
+
+export type { GroupInfoProps, GroupInfoResponseData };
